ARC-3272: Fix rounding issue since it would allow duplicates

+++ src/modules/material-requests/services/material-requests.service.ts
@@ -488,7 +488,7 @@ export class MaterialRequestsService {
 				}
 
 				if (keyValue.key === 'startTime' || keyValue.key === 'endTime') {
-					const parsed = Number.parseInt(keyValue.value);
+					const parsed = Number.parseFloat(keyValue.value);
 					reuseForm[keyValue.key] = Number.isNaN(parsed) ? null : parsed;
 				} else {
 					reuseForm[keyValue.key] = keyValue.value;
